Reddis added

+++ script/entrypoint.sh
@@ -5,12 +5,14 @@ TRY_LOOP="${TRY_LOOP:-10}"
 POSTGRES_HOST="${POSTGRES_HOST:-postgres}"
 POSTGRES_PORT=5432
 POSTGRES_CREDS="${POSTGRES_CREDS:-airflow:airflow}"
+REDIS_HOST="${REDIS_HOST:-redis}"
+REDIS_PORT=6379
 RABBITMQ_HOST="${RABBITMQ_HOST:-rabbitmq}"
 RABBITMQ_CREDS="${RABBITMQ_CREDS:-airflow:airflow}"
 RABBITMQ_MANAGEMENT_PORT=15672
 FLOWER_URL_PREFIX="${FLOWER_URL_PREFIX:-}"
 AIRFLOW_URL_PREFIX="${AIRFLOW_URL_PREFIX:-}"
-LOAD_DAGS_EXAMPLES="${LOAD_DAGS_EXAMPLES:-true}"
+LOAD_DAGS_EXAMPLES="${LOAD_DAGS_EXAMPLES:-False}"
 GIT_SYNC_REPO="${GIT_SYNC_REPO:-}"
 
 if [ -z $FERNET_KEY ]; then
@@ -28,6 +30,7 @@ sed -i "s/{{ FERNET_KEY }}/${FERNET_KEY}/" $AIRFLOW_HOME/airflow.cfg
 sed -i "s/{{ POSTGRES_HOST }}/${POSTGRES_HOST}/" $AIRFLOW_HOME/airflow.cfg
 sed -i "s/{{ POSTGRES_CREDS }}/${POSTGRES_CREDS}/" $AIRFLOW_HOME/airflow.cfg
 sed -i "s/{{ RABBITMQ_HOST }}/${RABBITMQ_HOST}/" $AIRFLOW_HOME/airflow.cfg
+sed -i "s/{{ REDIS_HOST }}/${REDIS_HOST}/" $AIRFLOW_HOME/airflow.cfg
 sed -i "s/{{ RABBITMQ_CREDS }}/${RABBITMQ_CREDS}/" $AIRFLOW_HOME/airflow.cfg
 sed -i "s/{{ LOAD_DAGS_EXAMPLES }}/${LOAD_DAGS_EXAMPLES}/" $AIRFLOW_HOME/airflow.cfg
 sed -i "s#{{ FLOWER_URL_PREFIX }}#${FLOWER_URL_PREFIX}#" $AIRFLOW_HOME/airflow.cfg
@@ -47,18 +50,38 @@ if [ "$1" = "webserver" ] || [ "$1" = "worker" ] || [ "$1" = "scheduler" ] || [ 
   done
 fi
 
+# wait for redis
+if [ "$1" = "webserver" ] || [ "$1" = "worker" ] || [ "$1" = "scheduler" ] || [ "$1" = "flower" ] ; then
+  j=0
+  while ! curl http://$REDIS_HOST:$REDIS_PORT/SET/hello/world |grep '200 OK'; do
+    j=`expr $j + 1`
+    if [ $j -ge $TRY_LOOP ]; then
+      echo "$(date) - $REDIS_HOST still not reachable, giving up"
+      exit 1
+    fi
+    echo "$(date) - waiting for REDIS... $j/$TRY_LOOP"
+	
+    sleep 5
+  done
+fi
+
+
+
 # wait for postgres
 if [ "$1" = "webserver" ] || [ "$1" = "worker" ] || [ "$1" = "scheduler" ] ; then
   i=0
   while ! nc $POSTGRES_HOST $POSTGRES_PORT >/dev/null 2>&1 < /dev/null; do
     i=`expr $i + 1`
-    if [ $i -ge $TRY_LOOP ]; then
+   =p]- if [ $i -ge $TRY_LOOP ]; then
       echo "$(date) - ${POSTGRES_HOST}:${POSTGRES_PORT} still not reachable, giving up"
       exit 1
     fi
     echo "$(date) - waiting for ${POSTGRES_HOST}:${POSTGRES_PORT}... $i/$TRY_LOOP"
     sleep 5
   done
+  
+
+  
   # TODO: move to a Helm hook
   #   https://github.com/kubernetes/helm/blob/master/docs/charts_hooks.md
   if [ "$1" = "webserver" ]; then
@@ -67,12 +90,15 @@ if [ "$1" = "webserver" ] || [ "$1" = "worker" ] || [ "$1" = "scheduler" ] ; the
   fi
 fi
 
-if [ ! -z $GIT_SYNC_REPO ]; then
+if [ "$1" = "worker" ] || [ "$1" = "scheduler" ] ; then
+  if [ ! -z $GIT_SYNC_REPO ]; then
     mkdir -p $AIRFLOW_HOME/dags
     # remove possible embedded dags to avoid conflicts
     rm -rf $AIRFLOW_HOME/dags/*
     echo "Executing background task git-sync on repo $GIT_SYNC_REPO"
     $AIRFLOW_HOME/git-sync --dest $AIRFLOW_HOME/dags --force &
+  fi
 fi
 
 $CMD "$@"
+
